Restructure, NodeGUI works*
*sorta
When building `npm build` or `npm start` it likes to create two .node files. One is not a proper executable and just points to the other, which is.. Solution is to delete the fake node and rename the real one to the fake one. 🤷‍♂️

run with `npm run` and do changes, then run `qode /dist/index.js`