#44: Fix ChaCha20-Poly1305, finalized class
ChaCha20-Poly1305 fixed, issue with the IvParameterSpec .. cannot use ChaCha20, but can use ChaCha20-Poly1305. I didn't make the rules. (see comments inside file).

Fix the HDKF key function, was converting the secure key to a string (?) then to bytes. This was incorrect, the secure key was already encoded correctly as bytes, running to string likely returned the hash for the object NOT the value. This led to incorrect key generation which led to tag mismatches when decrypting server messages.