#62: Gets pairing working
If the pair key does not exists, client will attempt to pair. Likewise, server ignores requests until a pair request comes in.

Pair requests tells server to generate a pairId and pairKey (see doc), then sends that back to the client. No fancy logic.

After pair, both server and client retain the configuration for the other. In server, if a client is not paired when shutting down, we delete the config file. On client we don't do this currently, so there could be extra servers laying around. If a pair error occurs on client, we do delete the server.