Adds NeptuneRunner (launcher for Neptune)
NeptuneRunner is the easiest way to implement notification handling via C# and easily connect the two applications.

NeptuneRunner starts the qode process as a child, pipes the STDIN, STDOUT, and STDERR from qode to NeptuneRunner (proxying the console).
NeptuneRunner also hosts a named pipe for NeptuneServer and NeptuneRunner to talk over. This will be used for sending notification data.

Additionally, NeptuneRunner (via user32.dll) is able to modify the taskbar details for NeptuneServer. That is, NeptuneServer now appears in the taskbar as "Neptune Server" rather than "Qode ..." in addition to being able to be pinned and launched from the taskbar correctly now.