Removes save button, auto update Server UI, notification fixes
Try to address notification issues, not 100%.
Using "*Compat" where possible as opposed to the UWP notification library to help address some issues, there are still major issue with the notification system. I don't believe these may get solved but here we are.

I believe people are not using the save button and instead expected the checkboxes to (when checked/unchecked) to get saved. This was not the case, the save button was meant to SAVE the checkbox states to the client configuration. Okay, so removed the save button entirely and now when the checkbox states is modified the configuration is updated and saved.

Added events to the server, the UI (MainWindow) now updates when the client configuration changes (battery info, configuration sync).
The Client's Server settings page now also updates the checkbox states when the configuration is updated (receiving the server's configuration).
This means unchecking "enable file sharing" on one device -> sends the new configuration to the other device -> other device updates their configuration data -> fires event -> UI gets updated to reflect the change.

Fix client getting HTML clipboard data, was not pulling the plaintext data as well.

A bunch of various fixes from the "fix JSDoc" commit (6d26a363131f5910f1279072ad0a8a12baf200da), because that borked a lot.